[add] プロジェクト削除時に確認ダイアログを出す

+++ src/renderer/component/presentational/taskManager/projectList/project.tsx
@@ -11,7 +11,9 @@ export interface Props{
 
 export class Project extends React.Component<Props, any> {
     onDeleteButtonClick(event: React.MouseEvent) {
-        this.props.onDeleteButtonClick()
+        if(confirm("Delete [" + this.props.name + "] ?")){
+            this.props.onDeleteButtonClick()
+        }
         event.stopPropagation()
     }
